% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{load_database}
\alias{load_database}
\title{load_database}
\usage{
load_database(con)
}
\arguments{
\item{con}{a DBI connection to the database}
}
\description{
Load a Planner Arena benchmark database into a list of tibbles
}
